+++ http/src/main/kotlin/at/asitplus/wallet/backend/config/BackendConfigurationProperties.kt
@@ -22,21 +22,6 @@ data class BackendConfigurationProperties(
      * Configure details about revocation lists for Verifiable Credentials
      */
     val revocationList: RevocationListConfigurationProperties = RevocationListConfigurationProperties(),
-    /**
-     * Configure external service for getting OTT for EPrescription credentials
-     */
-    val eprescription: EPrescriptionConfigurationProperties? = null
-)
-
-data class EPrescriptionConfigurationProperties(
-    /**
-     * URL of the external service
-     */
-    val url: String = "https://example.com/",
-    /**
-     * Will be sent as header `X-API-Key` to the service at [url]
-     */
-    val apiKey: String = "",
 )
 
 data class CredentialConfigurationProperties(