+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 5.11.0:
  - Update to VC-K 5.11.0
  - Update dependencies, i.e. Spring Boot to 3.5.9
+ - Remove HealthID and ePrescription credentials
 
 5.10.0:
  - Update to VC-K 5.10.0